Scaling Ability Selection
We got the abilities (in sidebar) to scale with camera position and
size! (Good since we're going to have big maps)

Technical Stuff: This meant I had to change a couple thigs. The
instantiation of Ability Panel (its position) is now dependent on the
Ability Panel object it clones (the one in the map). Be sure to set its
position when making new maps! We can hard-code this in if we need to.